Move texture files to modules directory
