+++ io_scene_nif/modules/property/texture/__init__.py
