+++ io_scene_nif/modules/property/texture/texture_export.py
@@ -42,7 +42,7 @@ from pyffi.formats.nif import NifFormat
 from io_scene_nif.utility import nif_utils
 from io_scene_nif.utility.nif_logging import NifLog
 
-from io_scene_nif.texturesys.texture_writer import TextureWriter
+from io_scene_nif.modules.property.texturesys.texture_writer import TextureWriter
 from io_scene_nif.utility.nif_global import NifOp
 
 class TextureHelper():
