+++ io_scene_nif/modules/property/texture/texture_writer.py
